30_code/app/requirements.txt
python-dotenv
openai
markdown_pdf
flask
flask_cors
yfinance
google-generativeai
grip
